+++ pkg/ingester/ingester_test.go
@@ -30,7 +30,7 @@ import (
 	"github.com/failsafe-go/failsafe-go/circuitbreaker"
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
-	"github.com/gogo/status"
+	"github.com/grafana/dskit/grpcutil"
 	"github.com/grafana/dskit/kv"
 	dslog "github.com/grafana/dskit/log"
 	dskit_metrics "github.com/grafana/dskit/metrics"
@@ -2526,7 +2526,7 @@ func Test_Ingester_LabelNames(t *testing.T) {
 		i.utilizationBasedLimiter = &fakeUtilizationBasedLimiter{limitingReason: "cpu"}
 
 		_, err := i.LabelNames(ctx, &client.LabelNamesRequest{})
-		stat, ok := status.FromError(err)
+		stat, ok := grpcutil.ErrorToStatus(err)
 		require.True(t, ok)
 		require.Equal(t, codes.ResourceExhausted, stat.Code())
 		require.Equal(t, tooBusyErrorMsg, stat.Message())
@@ -2590,7 +2590,7 @@ func Test_Ingester_LabelValues(t *testing.T) {
 		i.utilizationBasedLimiter = &fakeUtilizationBasedLimiter{limitingReason: "cpu"}
 
 		_, err := i.LabelValues(ctx, &client.LabelValuesRequest{})
-		stat, ok := status.FromError(err)
+		stat, ok := grpcutil.ErrorToStatus(err)
 		require.True(t, ok)
 		require.Equal(t, codes.ResourceExhausted, stat.Code())
 		require.Equal(t, tooBusyErrorMsg, stat.Message())
@@ -2791,7 +2791,7 @@ func TestIngester_LabelNamesAndValues(t *testing.T) {
 		i.utilizationBasedLimiter = &fakeUtilizationBasedLimiter{limitingReason: "cpu"}
 
 		err := i.LabelNamesAndValues(&client.LabelNamesAndValuesRequest{}, nil)
-		stat, ok := status.FromError(err)
+		stat, ok := grpcutil.ErrorToStatus(err)
 		require.True(t, ok)
 		require.Equal(t, codes.ResourceExhausted, stat.Code())
 		require.Equal(t, tooBusyErrorMsg, stat.Message())
@@ -2911,7 +2911,7 @@ func TestIngester_LabelValuesCardinality(t *testing.T) {
 		i.utilizationBasedLimiter = &fakeUtilizationBasedLimiter{limitingReason: "cpu"}
 
 		err := i.LabelValuesCardinality(&client.LabelValuesCardinalityRequest{}, nil)
-		stat, ok := status.FromError(err)
+		stat, ok := grpcutil.ErrorToStatus(err)
 		require.True(t, ok)
 		require.Equal(t, codes.ResourceExhausted, stat.Code())
 		require.Equal(t, tooBusyErrorMsg, stat.Message())
@@ -3443,7 +3443,7 @@ func Test_Ingester_MetricsForLabelMatchers(t *testing.T) {
 		i.utilizationBasedLimiter = &fakeUtilizationBasedLimiter{limitingReason: "cpu"}
 
 		_, err := i.MetricsForLabelMatchers(ctx, &client.MetricsForLabelMatchersRequest{})
-		stat, ok := status.FromError(err)
+		stat, ok := grpcutil.ErrorToStatus(err)
 		require.True(t, ok)
 		require.Equal(t, codes.ResourceExhausted, stat.Code())
 		require.Equal(t, tooBusyErrorMsg, stat.Message())
@@ -3806,7 +3806,7 @@ func TestIngester_QueryStream(t *testing.T) {
 					it := chk.Iterator(nil)
 
 					require.Equal(t, chunkenc.ValHistogram, it.Next())
-					actualTs, actualHist := it.AtHistogram()
+					actualTs, actualHist := it.AtHistogram(nil)
 					require.Equal(t, int64(seriesID), actualTs)
 					require.Equal(t, util_test.GenerateTestHistogram(seriesID), actualHist)
 
@@ -3820,7 +3820,7 @@ func TestIngester_QueryStream(t *testing.T) {
 					it := chk.Iterator(nil)
 
 					require.Equal(t, chunkenc.ValFloatHistogram, it.Next())
-					actualTs, actualHist := it.AtFloatHistogram()
+					actualTs, actualHist := it.AtFloatHistogram(nil)
 					require.Equal(t, int64(seriesID), actualTs)
 					require.Equal(t, util_test.GenerateTestFloatHistogram(seriesID), actualHist)
 
@@ -3841,7 +3841,7 @@ func TestIngester_QueryStream(t *testing.T) {
 		i.utilizationBasedLimiter = &fakeUtilizationBasedLimiter{limitingReason: "cpu"}
 
 		err = i.QueryStream(&client.QueryRequest{}, nil)
-		stat, ok := status.FromError(err)
+		stat, ok := grpcutil.ErrorToStatus(err)
 		require.True(t, ok)
 		require.Equal(t, codes.ResourceExhausted, stat.Code())
 		require.Equal(t, tooBusyErrorMsg, stat.Message())
@@ -4300,7 +4300,7 @@ func TestIngester_QueryExemplars(t *testing.T) {
 		i.utilizationBasedLimiter = &fakeUtilizationBasedLimiter{limitingReason: "cpu"}
 
 		_, err := i.QueryExemplars(ctx, &client.ExemplarQueryRequest{})
-		stat, ok := status.FromError(err)
+		stat, ok := grpcutil.ErrorToStatus(err)
 		require.True(t, ok)
 		require.Equal(t, codes.ResourceExhausted, stat.Code())
 		require.Equal(t, tooBusyErrorMsg, stat.Message())
@@ -5615,7 +5615,7 @@ func Test_Ingester_UserStats(t *testing.T) {
 		i.utilizationBasedLimiter = &fakeUtilizationBasedLimiter{limitingReason: "cpu"}
 
 		_, err := i.UserStats(ctx, &client.UserStatsRequest{})
-		stat, ok := status.FromError(err)
+		stat, ok := grpcutil.ErrorToStatus(err)
 		require.True(t, ok)
 		require.Equal(t, codes.ResourceExhausted, stat.Code())
 		require.Equal(t, tooBusyErrorMsg, stat.Message())
@@ -6504,7 +6504,7 @@ func TestIngester_PushInstanceLimits(t *testing.T) {
 							assert.ErrorIs(t, err, testData.expectedErr)
 							var optional middleware.OptionalLogging
 							assert.ErrorAs(t, err, &optional)
-							s, ok := status.FromError(err)
+							s, ok := grpcutil.ErrorToStatus(err)
 							require.True(t, ok, "expected to be able to convert to gRPC status")
 							assert.Equal(t, codes.Unavailable, s.Code())
 						} else {
@@ -6686,20 +6686,20 @@ func TestIngester_PushInstanceLimitsWithCircuitBreaker(t *testing.T) {
 						// We push the last request 2 times.
 						// The first request causes an instance limit error,
 						// and which opens the circuit breaker, but no
-						// circuitbreaker.ErrCircuitBreakerOpen is returned.
+						// circuitbreaker.ErrOpen is returned.
 						_, err := client.Push(ctx, req)
 						require.Error(t, err)
-						require.NotErrorIs(t, err, circuitbreaker.ErrCircuitBreakerOpen)
-						s, ok := status.FromError(err)
+						require.NotErrorIs(t, err, circuitbreaker.ErrOpen)
+						s, ok := grpcutil.ErrorToStatus(err)
 						require.True(t, ok, "expected to be able to convert to gRPC status")
 						require.Equal(t, codes.Unavailable, s.Code())
 						require.ErrorContains(t, s.Err(), testData.expectedErr.Error())
 
-						// The second request causes a circuitbreaker.ErrCircuitBreakerOpen,
+						// The second request causes a circuitbreaker.ErrOpen,
 						// since the circuit breaker is already open.
 						_, err = client.Push(ctx, req)
 						require.Error(t, err)
-						require.ErrorIs(t, err, circuitbreaker.ErrCircuitBreakerOpen)
+						require.ErrorIs(t, err, circuitbreaker.ErrOpen)
 					}
 
 					// imitate time ticking between each push
@@ -6824,7 +6824,7 @@ func TestIngester_inflightPushRequests(t *testing.T) {
 		require.ErrorAs(t, err, &optional)
 		require.False(t, optional.ShouldLog(ctx, time.Duration(0)), "expected not to log via .ShouldLog()")
 
-		s, ok := status.FromError(err)
+		s, ok := grpcutil.ErrorToStatus(err)
 		require.True(t, ok, "expected to be able to convert to gRPC status")
 		require.Equal(t, codes.Unavailable, s.Code())
 
@@ -6931,7 +6931,7 @@ func TestIngester_inflightPushRequestsBytes(t *testing.T) {
 		require.ErrorAs(t, err, &optional)
 		require.False(t, optional.ShouldLog(ctx, time.Duration(0)), "expected not to log via .ShouldLog()")
 
-		s, ok := status.FromError(err)
+		s, ok := grpcutil.ErrorToStatus(err)
 		require.True(t, ok, "expected to be able to convert to gRPC status")
 		require.Equal(t, codes.Unavailable, s.Code())
 
@@ -9436,13 +9436,13 @@ func TestIngester_PushWithSampledErrors(t *testing.T) {
 					for i := 0; i < errorSampleRate-1; i++ {
 						_, err = client.Push(ctxs[0], req)
 						require.Error(t, err)
-						status, ok := status.FromError(err)
+						status, ok := grpcutil.ErrorToStatus(err)
 						require.True(t, ok)
 						require.ErrorContains(t, status.Err(), testData.expectedErrs[0].err.Error())
 					}
 					_, err = client.Push(ctxs[1], req)
 					require.Error(t, err)
-					status, ok := status.FromError(err)
+					status, ok := grpcutil.ErrorToStatus(err)
 					require.True(t, ok)
 					require.ErrorContains(t, status.Err(), testData.expectedErrs[1].err.Error())
 				}
@@ -9549,7 +9549,7 @@ func TestIngester_SampledUserLimitExceeded(t *testing.T) {
 		// Append 2 series first, expect max-series-per-user error.
 		_, err = client.Push(ctx, mimirpb.ToWriteRequest([][]mimirpb.LabelAdapter{metricLabelAdapters1, metricLabelAdapters2}, []mimirpb.Sample{sample2, sample3}, nil, nil, mimirpb.API))
 		require.Error(t, err)
-		status, ok := status.FromError(err)
+		status, ok := grpcutil.ErrorToStatus(err)
 		require.True(t, ok)
 		require.Errorf(t, expectedError, status.Message())
 	}
@@ -9653,7 +9653,7 @@ func TestIngester_SampledMetricLimitExceeded(t *testing.T) {
 		// Append 2 series first, expect max-series-per-user error.
 		_, err = client.Push(ctx, mimirpb.ToWriteRequest([][]mimirpb.LabelAdapter{metricLabelAdapters1, metricLabelAdapters2}, []mimirpb.Sample{sample2, sample3}, nil, nil, mimirpb.API))
 		require.Error(t, err)
-		status, ok := status.FromError(err)
+		status, ok := grpcutil.ErrorToStatus(err)
 		require.True(t, ok)
 		require.Errorf(t, expectedError, status.Message())
 	}