+++ pkg/mimir/modules.go
@@ -780,7 +780,7 @@ func (t *Mimir) initRulerStorage() (serv services.Service, err error) {
 
 func (t *Mimir) initRuler() (serv services.Service, err error) {
 	if t.RulerDirectStorage == nil {
-		level.Info(util_log.Logger).Log("msg", "The ruler storage has not been configured.  Not starting the ruler.")
+		level.Info(util_log.Logger).Log("msg", "The ruler storage has not been configured. Not starting the ruler.")
 		return nil, nil
 	}
 
@@ -883,10 +883,14 @@ func (t *Mimir) initRuler() (serv services.Service, err error) {
 }
 
 func (t *Mimir) initAlertManager() (serv services.Service, err error) {
-	// Initialize the compat package in classic mode. This has the same behavior as if
-	// the compat package was not initialized, but with debug logs when a configuration
-	// is loaded or a new configuration is submitted.
-	features, err := featurecontrol.NewFlags(util_log.Logger, featurecontrol.FeatureClassicMode)
+	mode := featurecontrol.FeatureClassicMode
+	if t.Cfg.Alertmanager.UTF8StrictMode {
+		level.Info(util_log.Logger).Log("msg", "Starting Alertmanager in UTF-8 strict mode")
+		mode = featurecontrol.FeatureUTF8StrictMode
+	} else {
+		level.Info(util_log.Logger).Log("msg", "Starting Alertmanager in classic mode")
+	}
+	features, err := featurecontrol.NewFlags(util_log.Logger, mode)
 	util_log.CheckFatal("initializing Alertmanager feature flags", err)
 	compat.InitFromFlags(util_log.Logger, compat.RegisteredMetrics, features)