+++ pkg/querier/iterators/chunk_iterator.go
@@ -24,11 +24,13 @@ type chunkIterator struct {
 	valType chunkenc.ValueType
 	// AtT() is called often in the heap code, so caching its result seems like
 	// a good idea.
-	cacheValid           bool
-	cachedTime           int64
-	cachedValue          float64
-	cachedHistogram      *histogram.Histogram
-	cachedFloatHistogram *histogram.FloatHistogram
+	cachedValueValid          bool
+	cachedHistogramValid      bool
+	cachedFloatHistogramValid bool
+	cachedTime                int64
+	cachedValue               float64
+	cachedHistogram           *histogram.Histogram
+	cachedFloatHistogram      *histogram.FloatHistogram
 }
 
 // Seek advances the iterator forward to the value at or after
@@ -53,61 +55,56 @@ func (i *chunkIterator) At() (int64, float64) {
 		panic(fmt.Errorf("chunkIterator: calling At when chunk is of different type %v", i.valType))
 	}
 
-	if i.cacheValid {
+	if i.cachedValueValid {
 		return i.cachedTime, i.cachedValue
 	}
 
 	v := i.it.Value()
 	i.cachedTime, i.cachedValue = int64(v.Timestamp), float64(v.Value)
-	i.cacheValid = true
+	i.cachedValueValid = true
 	return i.cachedTime, i.cachedValue
 }
 
-func (i *chunkIterator) AtHistogram() (int64, *histogram.Histogram) {
+func (i *chunkIterator) AtHistogram(h *histogram.Histogram) (int64, *histogram.Histogram) {
 	if i.valType != chunkenc.ValHistogram {
 		panic(fmt.Errorf("chunkIterator: calling AtHistogram when chunk is of different type %v", i.valType))
 	}
 
-	if i.cacheValid {
+	if i.cachedHistogramValid {
 		return i.cachedTime, i.cachedHistogram
 	}
 
-	i.cachedTime, i.cachedHistogram = i.it.AtHistogram()
-	i.cacheValid = true
+	i.cachedTime, i.cachedHistogram = i.it.AtHistogram(h)
+	i.cachedHistogramValid = true
 	return i.cachedTime, i.cachedHistogram
 }
 
-func (i *chunkIterator) AtFloatHistogram() (int64, *histogram.FloatHistogram) {
+func (i *chunkIterator) AtFloatHistogram(fh *histogram.FloatHistogram) (int64, *histogram.FloatHistogram) {
 	if i.valType != chunkenc.ValHistogram && i.valType != chunkenc.ValFloatHistogram {
 		panic(fmt.Errorf("chunkIterator: calling AtFloatHistogram when chunk is of type %v", i.valType))
 	}
 
-	if i.cacheValid && i.cachedFloatHistogram != nil {
+	if i.cachedFloatHistogramValid {
 		return i.cachedTime, i.cachedFloatHistogram
 	}
 
-	var (
-		t  int64
-		fh *histogram.FloatHistogram
-	)
-
+	var t int64
 	if i.valType == chunkenc.ValHistogram {
-		var h *histogram.Histogram
-		t, h = i.AtHistogram()
-		fh = h.ToFloat(nil)
+		t, i.cachedHistogram = i.AtHistogram(i.cachedHistogram)
+		fh = i.cachedHistogram.ToFloat(fh)
 	} else {
-		t, fh = i.it.AtFloatHistogram()
+		t, fh = i.it.AtFloatHistogram(fh)
 	}
 
 	i.cachedTime = t
 	i.cachedFloatHistogram = fh
-	i.cacheValid = true
+	i.cachedFloatHistogramValid = true
 
 	return i.cachedTime, i.cachedFloatHistogram
 }
 
 func (i *chunkIterator) AtT() int64 {
-	if i.cacheValid {
+	if i.cachedValueValid || i.cachedHistogramValid || i.cachedFloatHistogramValid {
 		return i.cachedTime
 	}
 	switch i.valType {
@@ -115,10 +112,10 @@ func (i *chunkIterator) AtT() int64 {
 		t, _ := i.At()
 		return t
 	case chunkenc.ValHistogram:
-		t, _ := i.AtHistogram()
+		t, _ := i.AtHistogram(nil)
 		return t
 	case chunkenc.ValFloatHistogram:
-		t, _ := i.AtFloatHistogram()
+		t, _ := i.AtFloatHistogram(nil)
 		return t
 	default:
 		panic(fmt.Errorf("chunkIterator: calling AtT with unknown chunk encoding %v", i.valType))
@@ -143,7 +140,7 @@ func (i *chunkIterator) Err() error {
 }
 
 func (i *chunkIterator) invalidateCache() {
-	i.cacheValid = false
-	i.cachedHistogram = nil
-	i.cachedFloatHistogram = nil
+	i.cachedValueValid = false
+	i.cachedHistogramValid = false
+	i.cachedFloatHistogramValid = false
 }